requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
dataclasses-json==0.5.7
decorator==5.1.1
docx2txt==0.8
faiss-cpu==1.7.4
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.169
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.3
openai==0.27.6
packaging==23.1
pdfminer.six==20221105
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
slack-bolt==1.18.0
slack-sdk==3.21.3
SQLAlchemy==2.0.13
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
validators==0.20.0
weaviate-client==3.18.0
yarl==1.9.2
